+++ react/src/Components.js
@@ -6,8 +6,6 @@ import Time from './utils/Time';
 
 export const Link = ({to, env = true, tip = null, bold = true, children}) => {
     return <ReactLink className={tip ? "tool-tip" : ""} data-text={tip} to={Client.Path(to, env ? env : null)} style={{color:"inherit",fontWeight:bold ? "bold" : "inherit"}}>{children}</ReactLink>
-    return tip ? <ReactLink className={tip ? "tool-tip" : ""} data-text={tip} to={Client.Path(to, env ? env : null)} style={{color:"inherit",fontWeight:bold ? "bold" : "inherit"}}>{children}</ReactLink>
-               : <ReactLink to={Client.Path(to, env ? env : null)} style={{color:"inherit",fontWeight:bold ? "bold" : "inherit"}}>{children}</ReactLink>
 }
 
 export const HorizontalLine = ({top = "0", bottom = "0", thick = false, color = null }) => {
@@ -20,7 +18,7 @@ export const LoggedInUser = ({ link = undefined}) => {
     if (link === "user") {
         link = "/users/" + Auth.LoggedInUser();
     }
-    else if (link == undefined) {
+    else if (link === undefined) {
         link = "/login";
     }
     return <>
@@ -31,12 +29,12 @@ export const LoggedInUser = ({ link = undefined}) => {
         </>}
         { Auth.LoggedInViaGoogle() ? <>
             <span className="tool-tip" data-text="Google Authentication">
-                <img title="Via Google" style={{marginLeft:"9px",marginRight:"0",marginBottom:"2px"}} src={Image.GoogleLoginLogo()} height="15" />
+                <img alt="google" title="Via Google" style={{marginLeft:"9px",marginRight:"0",marginBottom:"2px"}} src={Image.GoogleLoginLogo()} height="15" />
             </span>
         </>:<>
             { Auth.LoggedInViaGitHub() && <>
                 <span className="tool-tip" data-text="GitHub Authentication">
-                    <img title="Via GitHub" style={{marginLeft:"5px",marginRight:"-4px",marginBottom:"2px"}} src={Image.GitHubLoginLogo()} height="19" />
+                    <img alt="github" title="Via GitHub" style={{marginLeft:"5px",marginRight:"-4px",marginBottom:"2px"}} src={Image.GitHubLoginLogo()} height="19" />
                 </span>
             </>}
         </>}