+++ react/src/PagedTableComponent.js
@@ -1,15 +1,8 @@
 import { useEffect, useState } from 'react';
-import { Link, useParams, useSearchParams } from 'react-router-dom';
-import { RingSpinner } from './Spinners';
-import { useFetch } from './utils/Fetch';
-import Char from './utils/Char';
-import Client from './utils/Client';
+import { useSearchParams } from 'react-router-dom';
 import PaginationComponent from './PaginationComponent';
-import { PuffSpinner, BarSpinner } from './Spinners';
-import Server from './utils/Server';
 import Str from './utils/Str';
 import TableHead from './TableHead';
-import Time from './utils/Time';
 import Type from './utils/Type';
 
 const PagedTableComponent = ({columns, data, update, initialSort, children}) => {