+++ react/src/Styles.js
@@ -1,11 +1,8 @@
-// Styles ...
-
-// TODO
-// NOT IN USE YET
+// NOT YET IN IN WIDE USE
 // Starting to experiment with setting styles dynamically based on Fourfront vs. CGAP styles.
 
-const styles                          = getComputedStyle(document.documentElement)
 // TODO
+// const styles = getComputedStyle(document.documentElement)
 // For some reason reading these values, when NOT running in cross-origin local mode (UI on 3000 and API on 8000),
 // get garbled from the getPropertyValue calls below. Hardcode them for now ....
 // --foursight-fourfront-box-bg:#DDFFEE;
@@ -40,26 +37,6 @@ function SetFoursightCgapStyles() {
 // Also:
 // https://stackoverflow.com/questions/13375039/javascript-calculate-darker-colour
 //
-function oldLightenDarkenColor(col, amt) {
-    col = col.trim()
-    var usePound = false;
-    if (col[0] === "#") {
-        col = col.slice(1);
-        usePound = true;
-    }
-    var num = parseInt(col,16);
-    var r = (num >> 16) + amt;
-    if (r > 255) r = 255;
-    else if  (r < 0) r = 0;
-    var b = ((num >> 8) & 0x00FF) + amt;
-    if (b > 255) b = 255;
-    else if  (b < 0) b = 0;
-    var g = (num & 0x0000FF) + amt;
-    if (g > 255) g = 255;
-    else if (g < 0) g = 0;
-    col = (usePound?"#":"") + (g | (b << 8) | (r << 16)).toString(16);
-    return col;
-}
 function LightenDarkenColor(hexColor, magnitude) {
     hexColor = hexColor.trim()
     hexColor = hexColor.replace(`#`, ``);