+++ react/src/pages/UsersPage.js
@@ -1,13 +1,9 @@
-import { useEffect, useState } from 'react';
 import { Link, useParams, useSearchParams } from 'react-router-dom';
-import { RingSpinner } from '../Spinners';
 import { useFetch } from '../utils/Fetch';
 import Char from '../utils/Char';
 import Client from '../utils/Client';
-import { PuffSpinner, BarSpinner } from '../Spinners';
 import Server from '../utils/Server';
 import Str from '../utils/Str';
-import TableHead from '../TableHead';
 import PagedTableComponent from '../PagedTableComponent';
 import Time from '../utils/Time';
 import Type from '../utils/Type';